[Patch v5.4.6] Move sweep module and tune cooldown
