+++ tests/test_function_registry.py
@@ -29,13 +29,13 @@ FUNCTIONS_INFO = [
 
 
 
-    ("src/main.py", "parse_arguments", 1860),
-    ("src/main.py", "setup_output_directory", 1865),
-    ("src/main.py", "load_features_from_file", 1870),
-    ("src/main.py", "drop_nan_rows", 1875),
-    ("src/main.py", "convert_to_float32", 1880),
-    ("src/main.py", "run_initial_backtest", 1885),
-    ("src/main.py", "save_final_data", 1890),
+    ("src/main.py", "parse_arguments", 1851),
+    ("src/main.py", "setup_output_directory", 1856),
+    ("src/main.py", "load_features_from_file", 1861),
+    ("src/main.py", "drop_nan_rows", 1866),
+    ("src/main.py", "convert_to_float32", 1871),
+    ("src/main.py", "run_initial_backtest", 1876),
+    ("src/main.py", "save_final_data", 1881),
