+++ tests/test_soft_cooldown_logic.py
@@ -5,9 +5,9 @@ ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.insert(0, ROOT_DIR)
 sys.path.insert(0, os.path.join(ROOT_DIR, 'src'))
 
-SOFT_COOLDOWN_LOOKBACK = 10
-# [Patch v5.0.18] Updated threshold to match strategy defaults
-SOFT_COOLDOWN_LOSS_COUNT = 6
+SOFT_COOLDOWN_LOOKBACK = 15
+# [Patch v5.4.6] Adjust defaults for real market conditions
+SOFT_COOLDOWN_LOSS_COUNT = 2
 
 
 from cooldown_utils import is_soft_cooldown_triggered, step_soft_cooldown
